ActorLocationSenderComponentSystem
添加疑似缺少的代码

RouterComponentSystem

修正AwakeSystem类命名不规范问题

NumericComponentSystem

int 改 long 因为字典里存储的value为long型

+++ Unity/Assets/Scripts/Codes/Model/Share/Module/Numeric/NumericComponent.cs
@@ -24,7 +24,7 @@ namespace ET
 
         public static void Set(this NumericComponent self, int nt, float value)
         {
-            self[nt] = (int)(value * 10000);
+            self[nt] = (long)(value * 10000);
         }
 
         public static void Set(this NumericComponent self, int nt, int value)
